ReadMe
"This repository contains three folders: 'Simulation', 'Optimization', and 'Results'. In the 'Simulation' folder, you can simulate 

the Skeletal Human model equipped with a Gait Controller. For optimizing the gains of the gait controller and the Balance Distribution
 
Coefficient (BDC), please refer to the 'Optimization' folder. Additionally, the 'Results' folder contains the simulation results of the 

model for a six-cycle gait."